refactor: consolidate imports and move initLabel to init directory
