+++ src/init/index.ts
@@ -1,2 +1,3 @@
 export { addReleaseNoteTemplate } from "./addReleaseNoteTemplate.js";
 export { addPullRequestLabel } from "./addPullRequestLabel.js";
+export { initLabel } from "./initLabel.js";
